# api/generated/openapi.yaml
...
openapi: 3.1.0
info:
  title: AsiaTravelPlan Mobile API
  version: 2026-03-02.1
  summary: Contract for the in-house iPhone app and the AsiaTravelPlan backend.
  description: |-
    This contract is generated from the CUE abstract model (model/api, model/entities, model/enums, model/common).
    It describes the request and response payloads the iPhone app and frontend may rely on.

    Important:
    - The mobile app must not depend on backend storage details.
    - Breaking changes should increment the contract version and normally require a mobile app update.
    - Only these app roles are supported: atp_admin, atp_manager, atp_accountant, atp_staff, atp_tour_editor.
servers:
- url: https://api-staging.asiatravelplan.com
  description: Staging API
- url: http://localhost:8787
  description: Local development API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LanguageCode:
      type: string
      enum:
      - en
      - ar
      - fr
      - zh
      - ja
      - ko
      - vi
      - ms
      - de
      - es
      - it
      - ru
      - nl
      - pl
      - da
      - sv
      - 'no'
      x-enum-catalog:
      - code: en
      - code: ar
      - code: fr
      - code: zh
      - code: ja
      - code: ko
      - code: vi
      - code: ms
      - code: de
      - code: es
      - code: it
      - code: ru
      - code: nl
      - code: pl
      - code: da
      - code: sv
      - code: 'no'
      x-enum-options:
      - value: en
        label: en
      - value: ar
        label: ar
      - value: fr
        label: fr
      - value: zh
        label: zh
      - value: ja
        label: ja
      - value: ko
        label: ko
      - value: vi
        label: vi
      - value: ms
        label: ms
      - value: de
        label: de
      - value: es
        label: es
      - value: it
        label: it
      - value: ru
        label: ru
      - value: nl
        label: nl
      - value: pl
        label: pl
      - value: da
        label: da
      - value: sv
        label: sv
      - value: 'no'
        label: 'no'
...
    ATPCurrencyCode:
      type: string
      enum:
      - USD
      - EURO
      - VND
      - THB
      - CNY
      - JPY
      - KRW
      - RUB
      - PLN
      - DKK
      - SEK
      - NOK
      - AUD
      - GBP
      - NZD
      - ZAR
      x-enum-catalog:
      - code: USD
        symbol: "$"
        decimalPlaces: 2
      - code: EURO
        symbol: "€"
        decimalPlaces: 2
      - code: VND
        symbol: "₫"
        decimalPlaces: 0
      - code: THB
        symbol: "฿"
        decimalPlaces: 0
      - code: CNY
        symbol: CN¥
        decimalPlaces: 2
      - code: JPY
        symbol: "¥"
        decimalPlaces: 0
      - code: KRW
        symbol: "₩"
        decimalPlaces: 0
      - code: RUB
        symbol: "₽"
        decimalPlaces: 2
      - code: PLN
        symbol: zł
        decimalPlaces: 2
      - code: DKK
        symbol: kr
        decimalPlaces: 2
      - code: SEK
        symbol: kr
        decimalPlaces: 2
      - code: NOK
        symbol: kr
        decimalPlaces: 2
      - code: AUD
        symbol: A$
        decimalPlaces: 2
      - code: GBP
        symbol: "£"
        decimalPlaces: 2
      - code: NZD
        symbol: NZ$
        decimalPlaces: 2
      - code: ZAR
        symbol: R
        decimalPlaces: 2
      x-enum-options:
      - value: USD
        label: USD
      - value: EURO
        label: EURO
      - value: VND
        label: VND
      - value: THB
        label: THB
      - value: CNY
        label: CNY
      - value: JPY
        label: JPY
      - value: KRW
        label: KRW
      - value: RUB
        label: RUB
      - value: PLN
        label: PLN
      - value: DKK
        label: DKK
      - value: SEK
        label: SEK
      - value: NOK
        label: NOK
      - value: AUD
        label: AUD
      - value: GBP
        label: GBP
      - value: NZD
        label: NZD
      - value: ZAR
        label: ZAR
      x-currency-catalog:
      - code: USD
        symbol: "$"
        decimalPlaces: 2
      - code: EURO
        symbol: "€"
        decimalPlaces: 2
      - code: VND
        symbol: "₫"
        decimalPlaces: 0
      - code: THB
        symbol: "฿"
        decimalPlaces: 0
      - code: CNY
        symbol: CN¥
        decimalPlaces: 2
      - code: JPY
        symbol: "¥"
        decimalPlaces: 0
      - code: KRW
        symbol: "₩"
        decimalPlaces: 0
      - code: RUB
        symbol: "₽"
        decimalPlaces: 2
      - code: PLN
        symbol: zł
        decimalPlaces: 2
      - code: DKK
        symbol: kr
        decimalPlaces: 2
      - code: SEK
        symbol: kr
        decimalPlaces: 2
      - code: NOK
        symbol: kr
        decimalPlaces: 2
      - code: AUD
        symbol: A$
        decimalPlaces: 2
      - code: GBP
        symbol: "£"
        decimalPlaces: 2
      - code: NZD
        symbol: NZ$
        decimalPlaces: 2
      - code: ZAR
        symbol: R
        decimalPlaces: 2
...
    BookingSourceChannel:
      type: string
      enum:
      - other
      - website
      - email
      - whatsapp
      - facebook_messenger
      - google_maps
      - instagram
      - phone_call
      - walk_in
      - zalo
      x-enum-catalog:
      - code: other
      - code: website
      - code: email
      - code: whatsapp
      - code: facebook_messenger
      - code: google_maps
      - code: instagram
      - code: phone_call
      - code: walk_in
      - code: zalo
      x-enum-options:
      - value: other
        label: other
      - value: website
        label: website
      - value: email
        label: email
      - value: whatsapp
        label: whatsapp
      - value: facebook_messenger
        label: facebook_messenger
      - value: google_maps
        label: google_maps
      - value: instagram
        label: instagram
      - value: phone_call
        label: phone_call
      - value: walk_in
        label: walk_in
      - value: zalo
        label: zalo
    BookingReferralKind:
      type: string
      enum:
      - none
      - other_customer
      - b2b_partner
      - atp_staff
      x-enum-catalog:
      - code: none
      - code: other_customer
      - code: b2b_partner
      - code: atp_staff
      x-enum-options:
      - value: none
        label: none
      - value: other_customer
        label: other_customer
      - value: b2b_partner
        label: b2b_partner
      - value: atp_staff
        label: atp_staff
...
    BookingPersonGender:
      type: string
      enum:
      - male
      - female
      - other
      - prefer_not_to_say
      x-enum-catalog:
      - code: male
      - code: female
      - code: other
      - code: prefer_not_to_say
      x-enum-options:
      - value: male
        label: male
      - value: female
        label: female
      - value: other
        label: other
      - value: prefer_not_to_say
        label: prefer_not_to_say
...
    AtpStaffProfile:
      type: object
      required:
      - username
      properties:
        username:
          type: string
        name:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        friendly_short_name:
          type: string
          nullable: true
        team_order:
          type: integer
          nullable: true
        picture_ref:
          type: string
          nullable: true
        languages:
          type: array
          items:
            "$ref": "#/components/schemas/LanguageCode"
        destinations:
          type: array
          items:
            "$ref": "#/components/schemas/CountryCode"
...
    BookingPersonAddress:
      type: object
      required: []
      properties:
        line_1:
          type: string
          nullable: true
        line_2:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state_region:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        country_code:
          "$ref": "#/components/schemas/CountryCode"
          nullable: true
...
    BookingPerson:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
        name:
          type: string
        photo_ref:
          type: string
          nullable: true
        emails:
          type: array
          items:
            type: string
            format: email
        phone_numbers:
          type: array
          items:
            type: string
        preferred_language:
          "$ref": "#/components/schemas/LanguageCode"
          nullable: true
        food_preferences:
          type: array
          items:
            type: string
        allergies:
          type: array
          items:
            type: string
        hotel_room_smoker:
          type: boolean
          nullable: true
        hotel_room_sharing_ok:
          type: boolean
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        gender:
          "$ref": "#/components/schemas/BookingPersonGender"
          nullable: true
        nationality:
          "$ref": "#/components/schemas/CountryCode"
          nullable: true
        address:
          "$ref": "#/components/schemas/BookingPersonAddress"
          nullable: true
        roles:
          type: array
          items:
            "$ref": "#/components/schemas/BookingPersonRole"
        consents:
          type: array
          items:
            "$ref": "#/components/schemas/BookingPersonConsent"
        documents:
          type: array
          items:
            "$ref": "#/components/schemas/BookingPersonDocument"
        notes:
          type: string
          nullable: true
...
    BookingTravelPlan:
      type: object
      required: []
      properties:
        destinations:
          type: array
          items:
            "$ref": "#/components/schemas/CountryCode"
        days:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanDay"
        attachments:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanAttachment"
...
    BookingWebFormSubmission:
      type: object
      required: []
      properties:
        destinations:
          type: array
          items:
            "$ref": "#/components/schemas/CountryCode"
        travel_style:
          type: array
          items:
            type: string
        booking_name:
          type: string
          nullable: true
        tour_id:
          type: string
          nullable: true
        page_url:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        ip_country_guess:
          type: string
          nullable: true
        referrer:
          type: string
          nullable: true
        utm_source:
          type: string
          nullable: true
        utm_medium:
          type: string
          nullable: true
        utm_campaign:
          type: string
          nullable: true
        travel_month:
          type: string
          nullable: true
        number_of_travelers:
          type: integer
          nullable: true
        preferred_currency:
          "$ref": "#/components/schemas/ATPCurrencyCode"
          nullable: true
        travel_duration_days_min:
          type: integer
          nullable: true
        travel_duration_days_max:
          type: integer
          nullable: true
        name:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone_number:
          type: string
          nullable: true
        budget_lower_usd:
          type: integer
          nullable: true
        budget_upper_usd:
          type: integer
          nullable: true
        preferred_language:
          "$ref": "#/components/schemas/LanguageCode"
          nullable: true
        notes:
          type: string
          nullable: true
        submitted_at:
          type: string
          format: date-time
          nullable: true
...
    TranslationStatusSummary:
      type: object
      required:
      - lang
      - source_lang
      - status
      - stale
      - total_fields
      - translated_fields
      - missing_fields
      - has_source_content
      - has_target_content
      properties:
        lang:
          "$ref": "#/components/schemas/LanguageCode"
        source_lang:
          "$ref": "#/components/schemas/LanguageCode"
        status:
          type: string
        origin:
          type: string
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
        stale:
          type: boolean
        total_fields:
          type: integer
        translated_fields:
          type: integer
        missing_fields:
          type: integer
        has_source_content:
          type: boolean
        has_target_content:
          type: boolean
        source_hash:
          type: string
          nullable: true
...
    BookingOfferReadModel:
      type: object
      required:
      - currency
      - offer_detail_level_internal
      - offer_detail_level_visible
      - totals
      - visible_pricing
      - total_price_cents
      properties:
        status:
          type: string
          nullable: true
        currency:
          "$ref": "#/components/schemas/ATPCurrencyCode"
        offer_detail_level_internal:
          "$ref": "#/components/schemas/OfferDetailLevel"
        offer_detail_level_visible:
          "$ref": "#/components/schemas/OfferDetailLevel"
        category_rules:
          type: array
          items:
            "$ref": "#/components/schemas/BookingOfferCategoryRule"
        trip_price_internal:
          "$ref": "#/components/schemas/BookingOfferTripPriceInternal"
          nullable: true
        days_internal:
          type: array
          items:
            "$ref": "#/components/schemas/BookingOfferDayPriceInternal"
        additional_items:
          type: array
          items:
            "$ref": "#/components/schemas/BookingOfferAdditionalItem"
        discount:
          "$ref": "#/components/schemas/BookingOfferDiscount"
          nullable: true
        totals:
          "$ref": "#/components/schemas/BookingOfferTotals"
        quotation_summary:
          "$ref": "#/components/schemas/BookingOfferQuotationSummary"
          nullable: true
        payment_terms:
          "$ref": "#/components/schemas/BookingOfferPaymentTermsReadModel"
          nullable: true
        visible_pricing:
          "$ref": "#/components/schemas/BookingOfferVisiblePricingReadModel"
        total_price_cents:
          type: integer
    GeneratedBookingOfferReadModel:
      type: object
      required:
      - id
      - booking_id
      - version
      - filename
      - lang
      - created_at
      - currency
      - total_price_cents
      - offer
      - pdf_url
      properties:
        id:
          type: string
        booking_id:
          type: string
        version:
          type: integer
        filename:
          type: string
        lang:
          "$ref": "#/components/schemas/LanguageCode"
        comment:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          nullable: true
        currency:
          "$ref": "#/components/schemas/ATPCurrencyCode"
        total_price_cents:
          type: integer
        payment_terms:
          "$ref": "#/components/schemas/BookingOfferPaymentTermsReadModel"
          nullable: true
        offer:
          "$ref": "#/components/schemas/BookingOfferReadModel"
        travel_plan:
          "$ref": "#/components/schemas/BookingTravelPlan"
          nullable: true
        pdf_url:
          type: string
    BookingTravelPlanPdfReadModel:
      type: object
      required:
      - id
      - filename
      - page_count
      - created_at
      - sent_to_customer
      - pdf_url
      properties:
        id:
          type: string
        filename:
          type: string
        page_count:
          type: integer
        created_at:
          type: string
          format: date-time
        sent_to_customer:
          type: boolean
        comment:
          type: string
          nullable: true
        pdf_url:
          type: string
...
    BookingReadModel:
      type: object
      required:
      - id
      - offer
      - travel_plan_translation_status
      - generated_offer_email_enabled
      - translation_enabled
      - created_at
      - updated_at
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        assigned_keycloak_user_id:
          type: string
          nullable: true
        source_channel:
          "$ref": "#/components/schemas/BookingSourceChannel"
          nullable: true
        referral_kind:
          "$ref": "#/components/schemas/BookingReferralKind"
          nullable: true
        referral_label:
          type: string
          nullable: true
        referral_staff_user_id:
          type: string
          nullable: true
        assigned_keycloak_user_label:
          type: string
          nullable: true
        assigned_atp_staff:
          "$ref": "#/components/schemas/AtpStaffProfile"
          nullable: true
        core_revision:
          type: integer
          nullable: true
        notes_revision:
          type: integer
          nullable: true
        persons_revision:
          type: integer
          nullable: true
        travel_plan_revision:
          type: integer
          nullable: true
        offer_revision:
          type: integer
          nullable: true
        payment_documents_revision:
          type: integer
          nullable: true
        travel_styles:
          type: array
          items:
            type: string
        travel_start_day:
          type: string
          format: date
          nullable: true
        travel_end_day:
          type: string
          format: date
          nullable: true
        number_of_travelers:
          type: integer
          nullable: true
        preferred_currency:
          "$ref": "#/components/schemas/ATPCurrencyCode"
          nullable: true
        customer_language:
          "$ref": "#/components/schemas/LanguageCode"
          nullable: true
        notes:
          type: string
          nullable: true
        persons:
          type: array
          items:
            "$ref": "#/components/schemas/BookingPerson"
        travel_plan:
          "$ref": "#/components/schemas/BookingTravelPlan"
          nullable: true
        web_form_submission:
          "$ref": "#/components/schemas/BookingWebFormSubmission"
          nullable: true
        offer:
          "$ref": "#/components/schemas/BookingOfferReadModel"
        travel_plan_pdfs:
          type: array
          items:
            "$ref": "#/components/schemas/BookingTravelPlanPdfReadModel"
        generated_offers:
          type: array
          items:
            "$ref": "#/components/schemas/GeneratedBookingOfferReadModel"
        travel_plan_translation_status:
          "$ref": "#/components/schemas/TranslationStatusSummary"
        generated_offer_email_enabled:
          type: boolean
        translation_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BookingListFilters:
      type: object
      required: []
      properties:
        assigned_keycloak_user_id:
          type: string
          nullable: true
        search:
          type: string
          nullable: true
    BookingList:
      type: object
      required:
      - pagination
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/BookingReadModel"
        pagination:
          "$ref": "#/components/schemas/Pagination"
        filters:
          "$ref": "#/components/schemas/BookingListFilters"
          nullable: true
        sort:
          type: string
          nullable: true
...
    BookingDetail:
      type: object
      required:
      - booking
      properties:
        booking:
          "$ref": "#/components/schemas/BookingReadModel"
        unchanged:
          type: boolean
          nullable: true
...
    Pagination:
      type: object
      required:
      - page
      - page_size
      - total_items
      - total_pages
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        detail:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        request_id:
          type: string
          nullable: true
security:
- bearerAuth: []
paths:
...
  "/api/v1/bookings":
    get:
      operationId: bookings
      summary: Bookings
      tags:
      - Bookings
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingList"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    post:
      operationId: booking_create
      summary: Booking Create
      tags:
      - Bookings
      security:
      - bearerAuth: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BookingCreateRequest"
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingDetail"
        '201':
          description: Created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/BookingDetail"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...